+++ app/Http/Controllers/Webhook/IncomingWebhookController.php
@@ -5,6 +5,7 @@ namespace BristolSU\Module\Typeform\Http\Controllers\Webhook;
 use BristolSU\Module\Typeform\Http\Controllers\Controller;
 use BristolSU\Module\Typeform\Typeform\Handler\ResponseHandler;
 use BristolSU\Module\Typeform\Typeform\Handler\WebhookPayload;
+use BristolSU\Support\ModuleInstance\ModuleInstance;
 use Illuminate\Http\Request;
 
 class IncomingWebhookController extends Controller
@@ -13,7 +14,9 @@ class IncomingWebhookController extends Controller
     public function store(Request $request, ResponseHandler $handler)
     {
         $payload = new WebhookPayload($request->all());
-        $response = $handler->handle($payload);
+        if($payload->moduleInstanceId() === app(ModuleInstance::class)->id()) {
+            $response = $handler->handle($payload);
+        }
         
         return $response->load(['answers', 'answers.field']);
     }